aider: Changed the `my-8` class to `my-4` in the `src\components\projects\main.jsx` file to reduce the vertical margin around the projects section.
# Aider chat conversation:

USER: let's take a look at our projects component, what creates the space at the bottom

ASSISTANT: The space at the bottom of the projects component in the `src\components\projects\main.jsx` file is created by the combination of the `my-8` class on the `div` that wraps the children of the `Projects` component and the `pb-2` class on the `div` that contains the `h2` element with the text "PROJECTS". The `my-8` class adds margin to the top and bottom, and the `pb-2` class adds padding to the bottom of the element.

Here is the relevant section of the code:

```jsx
export default function Projects({ children }) {
    return (
        <div className="p-2">
            <div className="max-w-2xl mx-auto my-8">
                <div className="border-b-2 border-black pb-2">
                    <h2 className="text-2xl font-semibold">PROJECTS</h2>
                </div>
                {children}
            </div>
        </div>
    );
}
```

The `my-8` class on the second `div` is responsible for the vertical margin around the entire projects section, which includes the space at the bottom. If you want to adjust the space at the bottom, you could modify the `my-8` class to a different value or add an additional class to control the bottom margin specifically.

USER: let's try my-4

ASSISTANT: To adjust the vertical margin around the entire projects section to a smaller value, we will change the `my-8` class to `my-4` in the `src\components\projects\main.jsx` file. Here is the *SEARCH/REPLACE* block for this change:

src\components\projects\main.jsx
```jsx
<<<<<<< SEARCH
            <div className="max-w-2xl mx-auto my-8">
=======
            <div className="max-w-2xl mx-auto my-4">
>>>>>>> REPLACE
```

This change will reduce the top and bottom margin around the projects section, affecting the space at the bottom as well.